functions/fn-agent/requirements.txt
# fn-agent dependencies
# Server: nuclio_runner.py uses stdlib http.server — NO fastapi, NO uvicorn
#
# NOTE: llama-index, qdrant-client, and pydantic have tight interdependencies.
# Version pins are intentionally loose here so pip can resolve a compatible set.

# LlamaIndex core + integrations
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.3
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-qdrant>=0.2.0

# Embedding model runtime
sentence-transformers>=2.3.1
torch>=2.0.0

# Vector DB client — must be compatible with llama-index-vector-stores-qdrant
# (IDF_EMBEDDING_MODELS was added in qdrant-client 1.9.0)
qdrant-client>=1.9.0

# HTTP (used by Ollama SDK internally)
httpx>=0.27.0

# Pydantic — llama-index requires >=2.8.0 for pydantic.Secret
pydantic>=2.8.0
